ColPali/requirements.txt
requests
pdf2image
matplotlib
Pillow
transformers
torch
byaldi
qwen_vl_utils
accelerate
sentencepiece
tokenizers
gradio
openai>=1.0.0